Remove the ecnryption data offset from the header
Since the idea was to have encryption data and compression data before
the filestream, there was dedicated slots in the header to support
those.

This has been removed since the compression data should absolutely be
part of the encyrpted data of the file and the crypter abstraction can
just prepend data to the filestream if it needs to preserve data.